requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
langchain
langchain-community
langchain-chroma
langchain-huggingface
chromadb
pypdf
sentence-transformers
